Feat (articleController) : article 조회시 댓글 같이 보이도록 수정

+++ src/main/java/kr/codesqaud/cafe/controller/ArticleController.java
@@ -2,12 +2,22 @@ package kr.codesqaud.cafe.controller;
 
 import kr.codesqaud.cafe.domain.Article;
 import kr.codesqaud.cafe.domain.Member;
+import kr.codesqaud.cafe.dto.SessionUser;
+import kr.codesqaud.cafe.dto.answer.AnswerResponseDto;
+import kr.codesqaud.cafe.exception.ExceptionStatus;
+import kr.codesqaud.cafe.exception.InvalidAuthorityException;
+import kr.codesqaud.cafe.repository.AnswerRepository;
 import kr.codesqaud.cafe.repository.ArticleRepository;
 import kr.codesqaud.cafe.repository.MemberRepository;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+
+import javax.servlet.http.HttpSession;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 import static kr.codesqaud.cafe.constant.ConstUrl.REDIRECT_INDEX;
 
@@ -16,35 +26,58 @@ public class ArticleController {
 
     private final ArticleRepository articleRepository;
     private final MemberRepository memberRepository;
+    private final AnswerRepository answerRepository;
 
-    public ArticleController(ArticleRepository articleRepository, MemberRepository memberRepository) {
+    public ArticleController(ArticleRepository articleRepository, MemberRepository memberRepository, AnswerRepository answerRepository) {
         this.articleRepository = articleRepository;
         this.memberRepository = memberRepository;
+        this.answerRepository = answerRepository;
     }
 
     @PostMapping("/questions")
-    public String postQuestions(long writerId, String title, String contents) {
-        Member member = memberRepository.findById(writerId);
+    public String postQuestions(HttpSession httpSession, String title, String contents) {
+        SessionUser sessionUser = SessionUser.getSessionUser(httpSession);
+        Member member = memberRepository.findById(sessionUser.getId());
         Article article = new Article(member, title, contents);
         articleRepository.save(article);
         return REDIRECT_INDEX;
     }
 
     @DeleteMapping("/articles/{id}/delete")
-    public String deleteArticle(@PathVariable long id) {
+    public String deleteArticle(@PathVariable Long id, HttpSession httpSession) throws InvalidAuthorityException {
+        Article exArticle = articleRepository.findById(id);
+        SessionUser sessionUser = SessionUser.getSessionUser(httpSession);
+
+        if (!sessionUser.equals(exArticle.getWriterId())) {
+            throw new InvalidAuthorityException(ExceptionStatus.NO_SESSION_USER);
+        }
+
         articleRepository.delete(id);
         return REDIRECT_INDEX;
     }
 
     @GetMapping("/articles/{id}/update")
-    public String updateArticleForm(@PathVariable long id, Model model) {
-        model.addAttribute("article", articleRepository.findById(id));
+    public String updateArticleForm(@PathVariable long id, Model model, HttpSession httpSession) throws InvalidAuthorityException {
+        Article article = articleRepository.findById(id);
+        SessionUser sessionUser = SessionUser.getSessionUser(httpSession);
+
+        if (!sessionUser.equals(article.getWriterId())) {
+            throw new InvalidAuthorityException(ExceptionStatus.NO_SESSION_USER);
+        }
+
+        model.addAttribute("article", article);
         return "qna/updateForm";
     }
 
     @PutMapping("/articles/{id}/update")
-    public String updateArticle(Article newArticle, @PathVariable Long id, RedirectAttributes redirectAttributes) {
+    public String updateArticle(Article newArticle, @PathVariable Long id, RedirectAttributes redirectAttributes, HttpSession httpSession) throws InvalidAuthorityException {
         Article exArticle = articleRepository.findById(id);
+        SessionUser sessionUser = SessionUser.getSessionUser(httpSession);
+
+        if (!sessionUser.equals(exArticle.getWriterId())) {
+            throw new InvalidAuthorityException(ExceptionStatus.NO_SESSION_USER);
+        }
+
         articleRepository.update(exArticle, newArticle);
         redirectAttributes.addFlashAttribute("id", id);
         return "redirect:/articles/{id}";
@@ -53,6 +86,10 @@ public class ArticleController {
     @GetMapping("/articles/{id}")
     public String showArticle(@PathVariable Long id, Model model) {
         model.addAttribute("article", articleRepository.findById(id));
+
+        List<AnswerResponseDto> collect = answerRepository.findAll(id).stream().map(AnswerResponseDto::toDto).collect(Collectors.toList());
+        model.addAttribute("answers", collect);
+        model.addAttribute("answerSize", collect.size());
         return "qna/show";
     }
